refactor: consolidate recursive file traversal
